Implemented check of contains file in other events before delete

+++ app/src/main/java/com/appsverse/teethhistory/viewModels/EditEventViewModel.java
@@ -83,7 +83,7 @@ public class EditEventViewModel extends ViewModel {
         this.photosListForDeleting = photosListForDeleting;
     }
 
-    public void addListToPhotosListToDeleting(List<String> photosListForDeleting){
+    public void addListToPhotosListToDeleting(List<String> photosListForDeleting) {
         this.photosListForDeleting.addAll(photosListForDeleting);
     }
 
@@ -134,18 +134,41 @@ public class EditEventViewModel extends ViewModel {
         int amountOfNewPhotos = photosUri.size() - eventModelRealmList.size();
 
         for (int i = 1; i <= amountOfNewPhotos; i++) {
-            File file = new File(photosUri.get(photosUri.size()-i));
-            boolean isDeleted = file.delete();
-            if (isDeleted) {
-                Log.d(TAG, "file " + photosUri.get(photosUri.size()-i) + " deleted");
-            } else {
-                Log.d(TAG, "file " + photosUri.get(photosUri.size()-i) + " doesn't deleted");
+
+            String uri = photosUri.get(photosUri.size() - i);
+
+            if (!checkUriInOtherEvents(context, uri)) {
+                File file = new File(uri);
+                file.delete();
             }
         }
-
-        if (photosListForDeleting != null) photosListForDeleting.clear();
+         if (photosListForDeleting != null) photosListForDeleting.clear();
 
         setVisibilities(context);
+    }
+
+    private boolean checkUriInOtherEvents(Context context, String uri) {
+        UserModel userModel = realm.where(UserModel.class).equalTo("id", ((MainActivity) context).user_id).findFirst();
+        RealmList<ToothModel> toothModels = userModel.getToothModels();
+
+        int coincidenceCounter = 0;
+
+            for (ToothModel toothModel : toothModels) {
+
+                RealmList<EventModel> eventModels = toothModel.getEventModels();
+
+                for (EventModel eventModel : eventModels) {
+
+                    if (eventModel.getPhotosUri().contains(uri)) {
+                        coincidenceCounter++;
+                        break;
+                    }
+                }
+                if (coincidenceCounter > 1) return true;
+            }
+
+            return false;
+
     }
 
     public void onClickSaveButton(Event event, Context context) {
@@ -170,7 +193,7 @@ public class EditEventViewModel extends ViewModel {
         eventModel.setGuarantee(event.getGuarantee());
         eventModel.setNotes(event.getNotes());
 
-        RealmList<String>  eventModelRealmList = eventModel.getPhotosUri();
+        RealmList<String> eventModelRealmList = eventModel.getPhotosUri();
 
         eventModelRealmList.deleteAllFromRealm();
         eventModelRealmList.addAll(event.getPhotosUri());
@@ -179,7 +202,7 @@ public class EditEventViewModel extends ViewModel {
             returnToothModelStateIfLastActionFilled(toothModel);
         }
 
-        RealmResults<EventModel> eventModelsResults = toothModel.getEventModels().sort("date", Sort.DESCENDING,"id", Sort.DESCENDING);
+        RealmResults<EventModel> eventModelsResults = toothModel.getEventModels().sort("date", Sort.DESCENDING, "id", Sort.DESCENDING);
 
         if (eventModelsResults.get(0).getId() == eventModel.getId()
                 && !oldEventModelAction.equals(newEventModelAction)) {
@@ -231,9 +254,8 @@ public class EditEventViewModel extends ViewModel {
         mainActivity.binding.getTeethFormulaFragment().setTooth();
 
 
-
         if (photosListForDeleting != null) {
-            deleteSelectedPhotos(context, event);
+            deleteSelectedPhotos(context);
         }
 
     }
@@ -312,32 +334,15 @@ public class EditEventViewModel extends ViewModel {
         return userModel.getToothModels().where().equalTo("id", mainActivityViewModel.getChosenToothID()).findFirst();
     }
 
-    public void deleteSelectedPhotos(Context context, Event event) {
+    public void deleteSelectedPhotos(Context context) {
 
         //todo check and delete in not main thread?
 
-        UserModel userModel = realm.where(UserModel.class).equalTo("id", ((MainActivity) context).user_id).findFirst();
-        RealmList<ToothModel> toothModels = userModel.getToothModels();
-
-        int coincidenceCounter = 0;
         for (String uri : photosListForDeleting) {
-        for (ToothModel toothModel : toothModels) {
 
-            RealmList<EventModel> eventModels = toothModel.getEventModels();
-
-            for (EventModel eventModel : eventModels) {
-
-                    if (eventModel.getPhotosUri().contains(uri)) {
-                        coincidenceCounter++;
-                        break;
-                    }
-                }
-            if (coincidenceCounter>1) break;
-        }
-            if (coincidenceCounter<2) {
+            if (!checkUriInOtherEvents(context, uri)) {
                 File file = new File(uri);
-                boolean isDeleted = file.delete();
-                Log.d(TAG, "file deleted: " + isDeleted + " uri: " + uri);
+                file.delete();
             }
         }
         photosListForDeleting.clear();
